backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
contourpy==1.3.3
cycler==0.12.1
dj-database-url==3.1.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
fonttools==4.61.1
gunicorn==25.0.1
inflection==0.5.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.0.7
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
uritemplate==4.2.0
whitenoise==6.11.0
